requirements.txt
torch
torchvision
numpy
pandas
scipy
matplotlib
tensorflow==1.14
tensorboardX
PyConfigMaker
pillow
dill
mypy
